+++ quantrl/backends/torch.py
@@ -6,7 +6,7 @@
 __name__    = 'quantrl.backends.torch'
 __authors__ = ["Sampreet Kalita"]
 __created__ = "2024-03-10"
-__updated__ = "2024-03-22"
+__updated__ = "2024-03-23"
 
 # dependencies
 import numpy as np
@@ -149,3 +149,13 @@ class TorchBackend(BaseBackend):
     ) -> torch.Tensor:
         tensor[indices] = values
         return tensor
+
+    def iterate_i(self,
+        func,
+        iterations_i:int,
+        Y,
+        args:tuple=None
+    ):
+        for i in range(iterations_i):
+            Y = func(i, Y, args)
+        return Y