+++ quantrl/solvers/torch.py
@@ -6,14 +6,14 @@
 __name__    = 'quantrl.solvers.torch'
 __authors__ = ["Sampreet Kalita"]
 __created__ = "2024-03-10"
-__updated__ = "2024-03-20"
+__updated__ = "2024-03-23"
 
 # dependencies
 from torchdiffeq import odeint
 
 # quantrl modules
 from ..backends.torch import TorchBackend
-from .base import BaseIVPSolver, BaseIterativeSolver
+from .base import BaseIVPSolver
 
 # TODO: Implement interpolation
 
@@ -81,42 +81,3 @@ class TorchDiffEqIVPSolver(BaseIVPSolver):
         Y
     ):
         raise NotImplementedError
-
-class TorchIterativeSolver(BaseIterativeSolver):
-    """Iteratively solve using PyTorch.
-
-    Refer to :class:`quantrl.backends.base.BaseIterativeSolver` for its implementation.
-    """
-
-    def __init__(self,
-        func,
-        backend:TorchBackend=None
-    ):
-        # initialize BaseIVPSolver
-        super().__init__(
-            func=func,
-            backend=backend if backend is not None else TorchBackend(
-                precision='double'
-            )
-        )
-
-    def iterate(self,
-        y_0,
-        iterations:int,
-        args
-    ):
-        _Y = self.backend.empty(
-            shape=(iterations + 1, *self.backend.shape(
-                tensor=y_0
-            )),
-            dtype='real'
-        )
-        _Y[0] = y_0
-        for i in range(1, iterations + 1):
-            _Y = self.func(
-                i=i,
-                Y=_Y,
-                args=args
-            )
-
-        return _Y